jndmnfjmdl

+++ src/utils/cloudinary.js
@@ -10,19 +10,22 @@ cloudinary.config({
   api_secret: '3801qC7dEnP2dalcE8aXgdN8A8k' 
 });
 
-const uploadOnCloudinary = async (localFailePath) => {
+const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null
         //upload the file on cloudinary
-        const response = await cloudinary.v2.uploader.upload(localFilePath, {
+        const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         }) 
        
-       
+        fs.unlinkSync(localFilePath)
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+          }
+          // remove the locally saved temporary file as the upload operation got failed
         return null;
     }
 }
